Update README.md
Project 2: SQL

A crime has taken place and the detective needs your help. The detective gave you the crime scene report, but you somehow lost it. You vaguely remember that the crime was a murder that occurred sometime on Jan.15, 2018 and that it took place in SQL City. Start by retrieving the corresponding crime scene report from the police department’s database.

Follow these steps to solve this challenge:
- Download the `sql-murder-mystery.db` file 
- visit www.sqliteonline.com
- click on file, then open db and load in the database file you downloaded above
- write your SQL queries to see the different tables and the content
- use the schema diagram to navigate the different tables
- figure out who committed the crime with the details you remembered above
- create a word or txt document that contains your thought process, narrative and SQL codes written to arrive at the solution to the murder mystery
- submit a link to the word or txt document.